data_generator/requirements.txt
Faker
psycopg2-binary
tzdata
tqdm